# prog: optimize resourceCentric()