In practice, we need to try out many different corpus programs before we may find a matching resource. It's very inefficient to Clone() each of them.

This change gives a +76% speed improvement in the BenchmarkMutate() test.

```
          │ upstream      │              optimized                  │
          │    sec/op     │    sec/op     vs base                   │
Mutate-36     443.1µ ± 8%   103.1µ ± 31%  -76.73% (p=0.000 n=15+25)

          │ upstream      │              optimized                  │
          │    B/op       │    B/op     vs base                     │
Mutate-36    836.0Ki ± 1%   188.8Ki ± 0%  -77.42% (p=0.000 n=15+25)

          │ upstream      │              optimized                  │
          │  allocs/op    │      allocs/op     vs base              │
Mutate-36     2.643k ± 1%       1.778k ± 0%     -32.73% (n=15+25)
```